# apps/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: White House Village Manager API
  version: 0.1.0
  description: |
    Contrato da API de gestão da White House Village.

    Este arquivo é a **fonte da verdade**: é mergeado antes da implementação, e é
    contra ele que o painel e os testes são escritos. Pertence ao agente
    `tech-lead`; os demais propõem alterações, não editam.

    ## Campo desconhecido no corpo é erro, nunca silêncio

    Toda requisição com corpo **recusa** propriedade que o schema não declara,
    com `422 VALIDATION_ERROR` e o nome do campo em `details`. Não é purismo: um
    corpo com campo errado que responde `200` é a pior resposta possível — o
    cliente acha que gravou, o servidor não gravou nada, e não há erro em lugar
    nenhum para alguém investigar. Foi exatamente assim que
    `PATCH /reservations/{id}` com `{"guests": 6}` respondeu `200` e manteve
    `guests_count` em 4 (medido, 26/08/2026).

    A regra vale para `POST`, `PUT` e `PATCH`. Em `PATCH`, campo **ausente** não
    muda nada e campo com `null` limpa — ausência é silêncio combinado; nome
    desconhecido não é.

    ## Um dado, um nome

    Uma informação tem **um** nome em todo o contrato: o mesmo na entrada, na
    saída e na coluna do banco. Número de hóspedes é `guests_count` em pedido de
    orçamento, criação, edição, remarcação e resposta — não `guests` em metade
    delas.
servers:
  - url: http://localhost:8080/api/v1
    description: Desenvolvimento
...
paths:
...
  /rate-tables:
    get:
      tags: [Tarifário]
      summary: Lista as tabelas de tarifas
      description: |
        Tabela de tarifas é versionada por vigência (`valid_from`/`valid_to`).
        A vigente numa data é a de maior `valid_from` que cobre a data e está
        `active`; `valid_to` nulo é "sem fim".
      x-rbac: { recurso: settings, acao: ver }
      parameters:
        - { name: page,     in: query, schema: { type: integer, default: 1 } }
        - { name: per_page, in: query, schema: { type: integer, default: 25, maximum: 100 } }
        - { name: active,   in: query, schema: { type: boolean } }
        - name: on
          in: query
          schema: { type: string, format: date }
          description: Traz apenas a(s) tabela(s) vigente(s) na data informada.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/TabelaDeTarifas" }
                  - $ref: "#/components/schemas/Paginacao"
        "403": { $ref: "#/components/responses/Erro" }
    post:
      tags: [Tarifário]
      summary: Cria uma tabela de tarifas
      x-rbac: { recurso: settings, acao: criar }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TabelaDeTarifasEntrada" }
      responses:
        "201":
          description: Criada
          headers:
            Location:
              description: URI do recurso criado.
              schema: { type: string }
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/TabelaDeTarifas" }
        "409":
          description: "`CODE_IN_USE` — `name` já usado na propriedade. É a chave natural `UNIQUE (property_id, name)` que o seed usa para reencontrar a própria linha."
          $ref: "#/components/responses/Erro"
        "422": { $ref: "#/components/responses/Erro" }
...
  /rates:
    get:
      tags: [Tarifário]
      summary: Lista as tarifas (produto × tipo de data → valor)
      description: |
        `GET /rates?rate_table_id=...` reproduz a grade inteira da tabela — é o
        critério de aceite da spec §3 ("`GET /rates` reproduz a Tabela V1
        exatamente").
      x-rbac: { recurso: settings, acao: ver }
      parameters:
        - { name: page,     in: query, schema: { type: integer, default: 1 } }
        - { name: per_page, in: query, schema: { type: integer, default: 25, maximum: 100 } }
        - { name: rate_table_id, in: query, schema: { type: string, format: uuid } }
        - { name: unit_type_id,  in: query, schema: { type: string, format: uuid } }
        - { name: date_type,     in: query, schema: { $ref: "#/components/schemas/TipoDeData" } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Tarifa" }
                  - $ref: "#/components/schemas/Paginacao"
        "403": { $ref: "#/components/responses/Erro" }
    post:
      tags: [Tarifário]
      summary: Cria uma tarifa
      x-rbac: { recurso: settings, acao: criar }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TarifaEntrada" }
      responses:
        "201":
          description: Criada
          headers:
            Location:
              description: URI do recurso criado.
              schema: { type: string }
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Tarifa" }
        "409":
          description: "`CODE_IN_USE` — já há tarifa para esse `(rate_table_id, unit_type_id, date_type)`. Use `PATCH` ou `POST /rates/bulk`."
          $ref: "#/components/responses/Erro"
        "422": { $ref: "#/components/responses/Erro" }
...
  /holidays:
    get:
      tags: [Tarifário]
      summary: Lista os feriados
      description: Data em `holidays` classifica a noite como `feriado` (precedência 80).
      x-rbac: { recurso: settings, acao: ver }
      parameters:
        - { name: page,     in: query, schema: { type: integer, default: 1 } }
        - { name: per_page, in: query, schema: { type: integer, default: 25, maximum: 100 } }
        - { name: from,     in: query, schema: { type: string, format: date } }
        - { name: to,       in: query, schema: { type: string, format: date } }
        - { name: active,   in: query, schema: { type: boolean } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Feriado" }
                  - $ref: "#/components/schemas/Paginacao"
        "403": { $ref: "#/components/responses/Erro" }
    post:
      tags: [Tarifário]
      summary: Cria um feriado
      x-rbac: { recurso: settings, acao: criar }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FeriadoEntrada" }
      responses:
        "201":
          description: Criado
          headers:
            Location:
              description: URI do recurso criado.
              schema: { type: string }
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Feriado" }
        "409":
          description: "`CODE_IN_USE` — já existe feriado nessa data (`UNIQUE (property_id, date)`)."
          $ref: "#/components/responses/Erro"
        "422": { $ref: "#/components/responses/Erro" }
...
  /special-periods:
    get:
      tags: [Tarifário]
      summary: Lista os períodos especiais
      description: |
        Períodos **podem se sobrepor de propósito** (Réveillon dentro da alta
        temporada) — por isso a tabela não tem constraint de exclusão. Quem
        resolve o empate é a precedência (`date_type_rules`), que é dado.
      x-rbac: { recurso: settings, acao: ver }
      parameters:
        - { name: page,     in: query, schema: { type: integer, default: 1 } }
        - { name: per_page, in: query, schema: { type: integer, default: 25, maximum: 100 } }
        - { name: from,     in: query, schema: { type: string, format: date } }
        - { name: to,       in: query, schema: { type: string, format: date } }
        - { name: kind,     in: query, schema: { type: string, enum: [reveillon, carnaval, alta, evento] } }
        - { name: active,   in: query, schema: { type: boolean } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/PeriodoEspecial" }
                  - $ref: "#/components/schemas/Paginacao"
        "403": { $ref: "#/components/responses/Erro" }
    post:
      tags: [Tarifário]
      summary: Cria um período especial
      description: |
        `name` carrega o ano (`Réveillon 2026/2027`) porque é chave natural
        (`UNIQUE (property_id, name)`): o Réveillon do ano seguinte é linha nova,
        não edição desta.

        O intervalo é **inclusivo nas duas pontas** (`starts_on`..`ends_on`) —
        diferente da estadia, que é half-open. Período é uma faixa do calendário
        comercial, não uma ocupação: 28/12 a 02/01 classifica as seis datas,
        inclusive 02/01.
      x-rbac: { recurso: settings, acao: criar }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PeriodoEspecialEntrada" }
      responses:
        "201":
          description: Criado
          headers:
            Location:
              description: URI do recurso criado.
              schema: { type: string }
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/PeriodoEspecial" }
        "409":
          description: "`CODE_IN_USE` — `name` já usado na propriedade (chave natural do seed)."
          $ref: "#/components/responses/Erro"
        "422":
          description: "`VALIDATION_ERROR` — `ends_on` anterior a `starts_on`."
          $ref: "#/components/responses/Erro"
...
  /min-nights:
    get:
      tags: [Tarifário]
      summary: Lista as regras de estadia mínima
      description: |
        Mínimo por tipo de data, dentro de uma tabela de tarifas. Vale o **maior
        mínimo entre as noites da estadia** (spec §3) — quem aplica é o motor
        `booking.Build`, não o handler.
      x-rbac: { recurso: settings, acao: ver }
      parameters:
        - { name: page,     in: query, schema: { type: integer, default: 1 } }
        - { name: per_page, in: query, schema: { type: integer, default: 25, maximum: 100 } }
        - { name: rate_table_id, in: query, schema: { type: string, format: uuid } }
        - { name: date_type,     in: query, schema: { $ref: "#/components/schemas/TipoDeData" } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/MinimoDeNoites" }
                  - $ref: "#/components/schemas/Paginacao"
        "403": { $ref: "#/components/responses/Erro" }
    post:
      tags: [Tarifário]
      summary: Cria uma regra de estadia mínima
      x-rbac: { recurso: settings, acao: criar }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MinimoDeNoitesEntrada" }
      responses:
        "201":
          description: Criada
          headers:
            Location:
              description: URI do recurso criado.
              schema: { type: string }
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/MinimoDeNoites" }
        "409":
          description: "`CODE_IN_USE` — já há regra para esse `(rate_table_id, date_type)`."
          $ref: "#/components/responses/Erro"
        "422": { $ref: "#/components/responses/Erro" }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
...
  schemas:
    Paginacao:
      type: object
      properties:
        meta:
          type: object
          properties:
            page:        { type: integer }
            per_page:    { type: integer }
            total:       { type: integer, format: int64 }
            total_pages: { type: integer }
    TipoDeData:
      type: string
      enum: [normal, fds, feriado, alta, reveillon, carnaval]
      description: Resolvido por precedência — réveillon/carnaval > feriado > alta > fds > normal.
...
    TabelaDeTarifas:
      type: object
      properties:
        id:   { type: string, format: uuid }
        name: { type: string, example: Tabela Comercial V1, description: "Chave natural `UNIQUE (property_id, name)` — é por ela que o seed reencontra a própria linha." }
        valid_from: { type: string, format: date }
        valid_to:   { type: [string, "null"], format: date, description: "`null` = sem fim." }
        active:     { type: boolean }
        created_at: { type: string, format: date-time }
    TabelaDeTarifasEntrada:
      type: object
      required: [name, valid_from]
      properties:
        name:       { type: string, minLength: 2 }
        valid_from: { type: string, format: date }
        valid_to:   { type: [string, "null"], format: date }
        active:     { type: boolean, default: true }
...
    TarifaEntrada:
      type: object
      required: [rate_table_id, unit_type_id, date_type, amount_cents]
      properties:
        rate_table_id: { type: string, format: uuid }
        unit_type_id:  { type: string, format: uuid }
        date_type:     { $ref: "#/components/schemas/TipoDeData" }
        amount_cents:  { type: integer, format: int64, minimum: 1 }
...
security:
  - bearerAuth: []